Introduce Rolling Releases for libkdumpfile
We recently added release publishing with debian assets
for all 'release/*' branches. This commit enables the
same thing but for the 'master', '6.0/stage', and
'6.0/release' branches on every push.

One thing to note is that on every push we delete and
recreate the releases (including their related tags)
so as not to pollute our git and github metadata with
releases and tags that we probably won't be using.