DLPX-72774 Add libsnappy-dev as a runtime dependency for libkdumpfile
Projects like drgn that use pkg-config/pkgconf for detecting installed
library dependencies during the autoconf/automake step of their
compilation fail to detect libkdumpfile if libsnappy-dev is not
installed. For us this leads to drgn being compiled without support
for analyzing kdump-compressed crash dumps.

This patch essentially installs libsnappy-dev on every system where
we install the libkdumpfile package. This way when we install
libkdumpfile as a dependency when compiling drgn in linux-pkg, we
ensure that drgn correctly detects libkdumpfile and is build with
support for it.

A better solution to look into in the future which would require a
bit more research is to generate a separate libkdumpfile-dev package
(which itself has libsnappy-dev as a dependency) and list that new
package as a build-time dependency for drgn. For now though this
patch should effectively work around our issue.